+++ pages/_document.js
@@ -1,5 +1,4 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document'
-import Script from 'next/script'
 import BLOG from '@/blog.config'
 import CJK from '@/lib/cjk'
 class MyDocument extends Document {
@@ -86,53 +85,6 @@ class MyDocument extends Document {
 
           <link rel="icon" href="/favicon.ico" />
           <link rel="icon" href="/favicon.svg" type="image/svg+xml"></link>
-          {BLOG.analytics && BLOG.analytics.provider === 'ackee' && (
-            <Script
-              src={BLOG.analytics.ackeeConfig.tracker}
-              data-ackee-server={BLOG.analytics.ackeeConfig.dataAckeeServer}
-              data-ackee-domain-id={BLOG.analytics.ackeeConfig.domainId}
-            />
-          )}
-          {BLOG.autoCollapsedNavBar === true && (
-            <Script
-              dangerouslySetInnerHTML={{
-                __html: `
-              var windowTop=0;
-              function scrollTrigger(){
-                  let scrollS = window.scrollY;
-                  let nav = document.querySelector('.sticky-nav');
-                  if(scrollS >= windowTop){
-                      nav.style.opacity = 0;
-                      windowTop = scrollS;
-                  }else{
-                      nav.style.opacity = 1;
-                      windowTop = scrollS;
-                  }
-              };
-              window.addEventListener('scroll',scrollTrigger);
-          `
-              }}
-            />
-          )}
-          {BLOG.analytics && BLOG.analytics.provider === 'ga' && (
-            <>
-              <Script
-                src={`https://www.googletagmanager.com/gtag/js?id=${BLOG.analytics.gaConfig.measurementId}`}
-              />
-              <Script
-                dangerouslySetInnerHTML={{
-                  __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${BLOG.analytics.gaConfig.measurementId}', {
-              page_path: window.location.pathname,
-            });
-          `
-                }}
-              />
-            </>
-          )}
         </Head>
         <body className="bg-day dark:bg-night">
           <Main />
